Big Commit:  - Match JKRHeap  - Rejoink CodeWarrior  - Fix all compilation warnings, and enable  - Misc changes around

+++ decomp/SDK/src/dolphin/dvd/dvderror.c
@@ -2,9 +2,11 @@
 #include <dolphin/os.h>
 
 #include "../os/OSPrivate.h"
+#include "dvd_private.h"
 
-static u32 ErrorTable[] = { 0,          0x00023A00, 0x00062800, 0x00030200, 0x00031100, 0x00052000,
-                            0x00052001, 0x00052100, 0x00052400, 0x00052401, 0x00052402, 0x000B5A01,
+static u32 ErrorTable[] = { 0,          0x00023A00, 0x00062800, 0x00030200,
+                            0x00031100, 0x00052000, 0x00052001, 0x00052100,
+                            0x00052400, 0x00052401, 0x00052402, 0x000B5A01,
                             0x00056300, 0x00020401, 0x00020400, 0x00040800 };
 
 inline u8
@@ -16,7 +18,7 @@ ErrorCode2Num (u32 errorCode)
     {
         if (errorCode == ErrorTable[i])
         {
-            return i;
+            return (u8)i;
         }
     }
 
@@ -32,22 +34,22 @@ Convert (u32 error)
 
     if (error == 0x01234567)
     {
-        return -1;
+        return (u8)-1;
     }
     else if (error == 0x01234568)
     {
-        return -2;
+        return (u8)-2;
     }
 
     statusCode = (error >> 24) & 0xFF;
-    errorCode = error & 0x00FFFFFF;
-    errorNum = ErrorCode2Num (errorCode);
+    errorCode  = error & 0x00FFFFFF;
+    errorNum   = ErrorCode2Num (errorCode);
     if (statusCode >= 6)
     {
         statusCode = 6;
     }
 
-    return statusCode * 30 + errorNum;
+    return (u8)(statusCode * 30 + errorNum);
 }
 
 void
@@ -56,8 +58,8 @@ __DVDStoreErrorCode (u32 error)
     OSSramEx* sram;
     u8        num;
 
-    num = Convert (error);
-    sram = __OSLockSramEx();
+    num                = Convert (error);
+    sram               = __OSLockSramEx();
     sram->dvdErrorCode = num;
     __OSUnlockSramEx (TRUE);
 }
